mcp/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic>=2.7.4

# LangChain & LangGraph
langchain-core>=0.3.36
langchain-openai==0.2.2
langchain-tavily==0.2.0
langchain-mcp-adapters==0.1.2
langgraph==0.2.45
langgraph-checkpoint-sqlite>=2.0.0,<3.0.0

# MCP Server
fastmcp>=0.5.0

# Database
aiosqlite==0.20.0
